Moved the rest of the commands.
